chore: simplify api keys creation & editing
fixes FTD-2219
