+++ @3rdweb-sdk/react/hooks/useApi.ts
@@ -55,17 +55,20 @@ export interface ConfirmEmailInput {
   confirmationToken: string;
 }
 
+export type ApiKeyRecoverShareManagement = "AWS_MANAGED" | "USER_MANAGED";
+export type ApiKeyCustomAuthentication = {
+  jwksUri: string;
+  aud: string;
+};
+
 export type ApiKeyService = {
   id: string;
   name: string;
   targetAddresses: string[];
   actions: string[];
   // If updating here, need to update validation logic in `validation.ts` as well for recoveryShareManagement
-  recoveryShareManagement?: "AWS_MANAGED" | "USER_MANAGED";
-  customAuthentication?: {
-    jwksUri: string;
-    aud: string;
-  };
+  recoveryShareManagement?: ApiKeyRecoverShareManagement;
+  customAuthentication?: ApiKeyCustomAuthentication;
 };
 
 export type ApiKey = {
