+++ components/layout/app-shell/index.tsx
@@ -55,14 +55,15 @@ export const AppShell: ComponentWithChildren<AppShellProps> = ({
         position="sticky"
         top={0}
         zIndex="sticky"
-        boxShadow="md"
+        boxShadow="sm"
         w={{ md: hasSidebar ? SIDEBAR_WIDTH : "auto" }}
       >
         {" "}
       </GridItem>
       <GridItem
         minH={{ base: "100vh", md: "unset" }}
-        py={8}
+        pt={{ base: 6, md: 10 }}
+        pb={{ base: 6, md: 20 }}
         as="main"
         colSpan={{ base: 2, md: 1 }}
         rowSpan={1}
@@ -93,7 +94,8 @@ const AppHeader: React.FC = () => {
       rowSpan={1}
       background="backgroundHighlight"
       zIndex="sticky"
-      boxShadow="md"
+      boxShadow="sm"
+      pb={2}
     >
       <Container
         maxW="100%"
@@ -303,13 +305,13 @@ const AppFooter: React.FC = () => {
         category="footer"
         label="feedback"
       >
-        Feedback
+        <Text>Feedback</Text>
       </TrackedLink>
       <TrackedLink isExternal href="/privacy" category="footer" label="privacy">
-        Privacy Policy
+        <Text>Privacy Policy</Text>
       </TrackedLink>
       <TrackedLink isExternal href="/tos" category="footer" label="terms">
-        Terms of Service
+        <Text>Terms of Service</Text>
       </TrackedLink>
 
       <TrackedLink
@@ -319,7 +321,7 @@ const AppFooter: React.FC = () => {
         display={{ base: "none", md: "flex" }}
         label="gas-estimator"
       >
-        Gas Estimator
+        <Text>Gas Estimator</Text>
       </TrackedLink>
       <TrackedLink
         href="/chainlist"
@@ -328,9 +330,9 @@ const AppFooter: React.FC = () => {
         display={{ base: "none", md: "flex" }}
         label="chains"
       >
-        Chainlist
+        <Text>Chainlist</Text>
       </TrackedLink>
-      <Text alignSelf="center" order={{ base: 2, md: 0 }}>
+      <Text alignSelf="center" order={{ base: 2, md: 0 }} opacity={0.5}>
         thirdweb &copy; {new Date().getFullYear()}
       </Text>
     </GridItem>
