+++ components/settings/ApiKeys/RevokeButton.tsx
@@ -10,24 +10,24 @@ import {
 } from "@chakra-ui/react";
 import { useTrack } from "hooks/analytics/useTrack";
 import { useTxNotifications } from "hooks/useTxNotifications";
+import { useRouter } from "next/router";
 import { useRef } from "react";
 import { Button } from "tw-components";
 
 interface RevokeApiKeyButtonProps {
   id: string;
   name: string;
-  onRevoke: () => void;
 }
 
 export const RevokeApiKeyButton: React.FC<RevokeApiKeyButtonProps> = ({
   id,
   name,
-  onRevoke,
 }) => {
   const trackEvent = useTrack();
   const { isOpen, onOpen, onClose } = useDisclosure();
   const cancelRef = useRef<HTMLButtonElement>(null);
   const mutation = useRevokeApiKey();
+  const router = useRouter();
   const { onSuccess, onError } = useTxNotifications(
     "API Key revoked",
     "Failed to revoke an API Key",
@@ -45,7 +45,8 @@ export const RevokeApiKeyButton: React.FC<RevokeApiKeyButtonProps> = ({
     mutation.mutate(id, {
       onSuccess: () => {
         onSuccess();
-        onRevoke();
+        router.push("/dashboard/settings/api-keys");
+
         trackEvent({
           category: "api-keys",
           action: "revoke",
@@ -66,23 +67,24 @@ export const RevokeApiKeyButton: React.FC<RevokeApiKeyButtonProps> = ({
 
   return (
     <>
-      <Button colorScheme="red" onClick={onOpen}>
-        Revoke
+      <Button colorScheme="red" onClick={onOpen} variant="outline">
+        Delete key
       </Button>
 
       <AlertDialog
         isOpen={isOpen}
         leastDestructiveRef={cancelRef}
         onClose={onClose}
+        isCentered
       >
         <AlertDialogOverlay>
           <AlertDialogContent>
             <AlertDialogHeader fontSize="lg" fontWeight="bold">
-              Revoke {name}?
+              Delete {name}?
             </AlertDialogHeader>
 
             <AlertDialogBody>
-              Are you sure you want to revoke <strong>{name}</strong>? Any
+              Are you sure you want to delete <strong>{name}</strong>? Any
               integrations using this key will no longer be able to access
               thirdweb services.
             </AlertDialogBody>
@@ -97,7 +99,7 @@ export const RevokeApiKeyButton: React.FC<RevokeApiKeyButtonProps> = ({
                 ml={3}
                 isLoading={mutation.isLoading}
               >
-                I am sure
+                Delete
               </Button>
             </AlertDialogFooter>
           </AlertDialogContent>
