+++ package.json
@@ -88,6 +88,7 @@
     "@types/airtable": "^0.10.1",
     "@types/color": "^3.0.2",
     "@types/cors": "^2.8.14",
+    "@types/jest": "^29.5.6",
     "@types/nprogress": "^0.2.0",
     "@types/papaparse": "^5.3.10",
     "@types/pluralize": "^0.0.31",
