+++ pages/dashboard/engine/index.tsx
@@ -21,7 +21,7 @@ const EngineManage: ThirdwebNextPage = () => {
   const enginePermissions = useEnginePermissions(instanceUrl);
 
   return (
-    <Flex flexDir="column" gap={8} mt={{ base: 2, md: 6 }}>
+    <Flex flexDir="column" gap={8}>
       <Flex direction="column" gap={4}>
         <Flex direction="column" gap={2}>
           <Heading size="title.lg" as="h1">
