+++ pages/dashboard/settings/notifications.tsx
@@ -24,7 +24,7 @@ const SettingsNotificationsPage: ThirdwebNextPage = () => {
   const account = meQuery.data;
 
   return (
-    <Flex flexDir="column" gap={8} mt={{ base: 2, md: 6 }}>
+    <Flex flexDir="column" gap={8}>
       <Flex direction="column" gap={2}>
         <Heading size="title.lg" as="h1" h={10}>
           Notification Settings
