+++ pages/dashboard/settings/usage.tsx
@@ -27,7 +27,7 @@ const SettingsUsagePage: ThirdwebNextPage = () => {
   const account = meQuery.data;
 
   return (
-    <Flex flexDir="column" gap={8} mt={{ base: 2, md: 6 }}>
+    <Flex flexDir="column" gap={8}>
       <Flex direction="column">
         <Heading size="title.lg" as="h1">
           Usage
